test(contract): gate assignability fixture on declared surface change
The bidirectional assignability fixture compiled unconditionally in run.ts,
before the Vitest suite could apply the changeset-aware relaxation — so a
future breaking change that genuinely breaks base<->current assignability
would fail there regardless of its changeset. Skip that compile when the PR
declares a minor/major bump, matching the Vitest gate. Extract the bump
detection into shared.ts so run.ts and the suite share one implementation.
